Conflicts:
	public/js/constants.js
	public/js/custom.js
	public/js/views/DividendInvoice/list/ListView.js
	public/js/views/DividendPayments/list/ListView.js
	public/js/views/ExpensesInvoice/list/ListView.js
	public/js/views/Invoices/list/ListView.js
	public/js/views/JobPositions/list/ListView.js
	public/js/views/Leads/list/ListView.js
	public/js/views/Opportunities/list/ListView.js
	public/js/views/Orders/list/ListView.js
	public/js/views/Proforma/list/ListView.js
	public/js/views/Quotations/list/ListView.js
	public/js/views/customerPayments/list/ListView.js
	public/js/views/inventoryReport/list/ListView.js
	public/js/views/listViewBase.js
	public/js/views/salesInvoices/list/ListView.js
	public/js/views/salesQuotations/list/ListView.js
	public/js/views/trialBalance/list/ListView.js
	test/uiSpecs/views/balanceSheet.test.js
	test/uiSpecs/views/cashFlow.test.js
	test/uiSpecs/views/journalEntry.test.js
	test/uiSpecs/views/profitAndLoss.test.js
	test/uiSpecs/views/tasks.test.js
	test/uiSpecs/views/trialBalance.test.js

+++ public/js/views/Companies/form/FormView.js
@@ -209,7 +209,7 @@ define([
             );
             this.$el.find('.formLeftColumn').append(
                 new AttachView({
-                    model: this.formModel,
+                    model      : this.formModel,
                     contentType: 'Companies'
                 }).render().el
             );
@@ -380,6 +380,14 @@ define([
 
         deleteItems: function () {
             var mid = this.mId;
+            var answer;
+            
+            answer = confirm('Really DELETE item ?!');
+
+            if (answer === false) {
+                return false;
+            }
+            
             this.formModel.destroy({
                 headers: {
                     mid: mid
